Added Activity for handling notification PendingIntents
I've added the **IsItPoppinActivity** which is where the PendingIntent is sent when a notification is sent to the user asking if the current bar/night club is poppin'. The activity itself needs more work, it's currently barebones. 

After that, the LocationService needs further additions for verifying the user has been at the current location long enough to warrant asking, and that they haven't already replied.